create separate component for counting the response time
