+++ apps/factory/src/components/prompt_version.tsx
@@ -46,7 +46,7 @@ import {
 } from "~/generated/prisma-client-zod.ts";
 import PromotOutputLog from "./prompt_output_log";
 import { displayModes } from "~/validators/base";
-
+import Counter from "./counter_responsetime";
 function PromptVersion({
   ns,
   pp,
@@ -392,14 +392,3 @@ function PromptVersion({
 }
 
 export default PromptVersion;
-
-const Counter = () => {
-  const [seconds, setSeconds] = useState(0);
-  useEffect(() => {
-    const interval = setInterval(() => {
-      setSeconds(seconds + 1);
-    }, 1000);
-    return () => clearInterval(interval);
-  }, [seconds]);
-  return <>{seconds}</>;
-};
